Remove helper file, create folder and files for the classes
